DEMRUM-861: moving to more direct emission of data
